+++ src/main/kotlin/net/fallingangel/jimmerdto/psi/element/DTOImportedType.kt
@@ -1,9 +1,10 @@
 package net.fallingangel.jimmerdto.psi.element
 
+import com.intellij.psi.PsiElement
 import net.fallingangel.jimmerdto.psi.mixin.DTOElement
 
 interface DTOImportedType : DTOElement {
-    val type: String
+    val type: PsiElement
 
-    val alias: String?
+    val alias: PsiElement?
 }